requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
chromadb>=0.5.0
openai-whisper>=20240930
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
